modulo_P/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pydantic==2.5.0